Ping Pong 600 Started
So I started PP600. The idea is that I use a combination of moving the top element to the other stack as well as swapping the top two elements of one stack in order to move the top element of one stack to the bottom of the same stack. Then you can use one stack to hold both the string and the substitutions, and use the other stack to do temporary storage.